Methods for objects of class "profileCI"

Methods for objects of class "profileCI" returned from profileCI.

Usage

# S3 method for class 'profileCI'
print(x, ...)

# S3 method for class 'profileCI'
plot(x, parm = 1:nrow(x), add = TRUE, digits = 2, ...)

Arguments

x

An object inheriting from class profileCI", a result of a call to profileCI.

...

Further arguments. For print."profileCI" to pass arguments to print. For plot.profileCI to pass graphical parameters to graphics::plot to create the initial plot of the profile log-likelihood.

parm

A numeric or character scalar specifying the parameter for which a profile log-likelihood is plotted. Must be a single component consistent with the argument parm to profileCI.

add

A logical scalar. If add = TRUE then the plot is annotated with a horizontal line indicating the critical value for the profile log-likelihood used to calculate the confidence limits, vertical lines indicating the values of these limits and a legend stating the confidence interval.

digits

An integer. Passed to signif to round the confidence limits in the legend, if add = TRUE.

Value

print.profileCI: the argument x is returned, invisibly.

plot.profileCI: a numeric vector containing the confidence interval for the parameter chosen for the plot.

Details

print.profileCI. A numeric matrix with 2 columns giving the lower and upper confidence limits for the parameters specified by the argument parm in profileCI. These columns are labelled as (1-level)/2 and 1-(1-level)/2, expressed as a percentage, by default 2.5% and 97.5%.

plot.profileCI. A plot is produced of the profile log-likelihood for the parameter chosen by parm. Only the parameter values used to profile the log-likelihood in the call to profileCI are included, so if faster = TRUE was used then the plot will not be of a smooth curve but will be triangular in the middle.

Examples

See profileCI.

See also